Updating source URLs for north-america/us/nh (#2782)
Updating the following to reflect a change in the service:

* NH_2021_2022_6in_CIR.geojson

NH_2021_2022_6in_RGB.geojson was updated in a previous commit. The other
two still need to be updated.

NOTE:

Strangely, there is one difference between NH_2021_2022_6in_CIR.geojson
and NH_2021_2022_6in_RGB.geojson, which is that the latter has a `:None`
in the URL (Same place that the old URL scheme had a :2122CIR which
turns out needed to be removed). If I remove the `:None` from the RGB,
it seems to behave the same, so maybe we should just do that. But if I
add `:None` to the CIR, it becomes RGB.

BTW credit goes to someone in the OSM Slack, I didn't find this URL
myself.