Adjust end dates (#2778)
* 2024 imagery for the canton Zürich

Note: the imagery is currently missing a diagonal stripe that couldn't be recorded in 2024 due to the weather. For this reason I've currently set the "best" value to false.

* Remove best flag

See https://github.com/osmlab/editor-layer-index/pull/2716#issuecomment-2965082392

* Fixed start/end dates don't make really make sense

All three products are continously updated, the terrain ones in a six
year cycle the orthofoto faster. Making the start and end dates quite
meaningless outside of a rough indication from when the data is.

Maybe it would be a better idea to allow specification of an
"update_cycle" value, which would allow the user at least to determine
the maximum age of the imagery they are using.